moved testResults tsx/scss to same directory
