+++ src/components/TestResults/Statistics.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../styles/componentStyles/statistics.scss";
+import "./statistics.scss";
 import { useTestInformationContext } from "../../contexts/TestInformationContext";
 
 const Statistics = () => {
