+++ src/components/TestResults/leaderboard.scss
@@ -1,0 +1,31 @@
+.leaderboard-div {
+    margin-top: 1rem;
+    padding-left: 0.5rem;
+    padding-right: 0.5rem;
+    max-height: 20rem;
+    overflow-x: hidden;
+    overflow-y: auto;
+    scrollbar-gutter: stable both-edges;
+
+    table {
+        width: 100%;
+
+        thead {
+            border-bottom: 2px solid var(--primary-highlight-colour);
+        }
+
+        th,
+        td {
+            padding: 0.1rem 0.4rem;
+
+            &.right-align {
+                text-align: right;
+            }
+        }
+    }
+
+    .leaderboard-login-prompt {
+        text-align: center;
+        font-size: 30px;
+    }
+}
