+++ src/components/TestResults/statistics.scss
@@ -1,0 +1,40 @@
+@use "../../styles/constants";
+
+
+.test-results-statistics {
+    margin-top: 1rem;
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    grid-template-rows: repeat(2, 5rem);
+
+    .grid-item {
+        display: block;
+        text-align: center;
+
+        .score {
+            color: var(--primary-highlight-colour);
+            font-size: 2.3rem;
+        }
+
+        .label {
+            margin-top: -0.7rem;
+            color: var(--base-font-colour);
+        }
+
+        &.wpm {
+            display: grid;
+            grid-row: span 2;
+            grid-column: 3;
+            text-align: center;
+            font-size: 5.5rem;
+            color: var(--primary-highlight-colour);
+
+            .wpm-label {
+                color: var(--base-font-colour);
+                grid-area: 2/1;
+                margin-top: -2.2rem;
+                font-size: constants.$base-font-size;
+            }
+        }
+    }
+}
