+++ src/components/TestResults/wpm-graph.scss
@@ -1,0 +1,5 @@
+.results-chart {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
